tests and code coverage
