+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "package": "npx ncc build src/index.js -o dist --source-map --license licenses.txt",
-    "test": "jest"
+    "test": "npx jest",
+    "coverage": "npx make-coverage-badge --output-path ./badges/coverage.svg"
   },
   "repository": {
     "type": "git",
@@ -19,14 +20,42 @@
     "overwrite",
     "content"
   ],
+  "engines": {
+    "node": ">=22"
+  },
   "author": "Jeppe Ærenlund",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/Radioh/action-content-overwrite/issues"
   },
   "homepage": "https://github.com/Radioh/action-content-overwrite#readme",
+  "jest": {
+    "verbose": true,
+    "clearMocks": true,
+    "testEnvironment": "node",
+    "moduleFileExtensions": [
+      "js"
+    ],
+    "testMatch": [
+      "**/*.test.js"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/dist/"
+    ],
+    "coverageReporters": [
+      "json-summary",
+      "text",
+      "lcov"
+    ],
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "./src/**"
+    ]
+  },
   "dependencies": {
-    "@actions/core": "^1.11.1"
+    "@actions/core": "^1.11.1",
+    "make-coverage-badge": "^1.2.0"
   },
   "devDependencies": {
     "@vercel/ncc": "^0.38.2",
