+++ src/index.js
@@ -1,20 +1,6 @@
-const core = require("@actions/core");
-const fs = require("fs").promises;
-
-async function run() {
-  try {
-    const filePath = core.getInput("filePath", { required: true });
-    const content = core.getInput("content", { required: true });
-
-    if (!filePath || !content) {
-      throw new Error("filePath and content inputs are required");
-    }
-
-    await fs.writeFile(filePath, content, { encoding: "utf8", flag: "w" });
-    console.log(`Filled content into: ${filePath}`);
-  } catch (error) {
-    core.setFailed(error.message);
-  }
-}
+/**
+ * The entrypoint for the action.
+ */
+const { run } = require("./main");
 
 run();
